Deadlock

+++ files/js/src/main/scala/com/swoval/files/Observers.scala
@@ -1,5 +1,6 @@
 package com.swoval.files
 
+import com.swoval.concurrent.Lock
 import com.swoval.files.Directory.Entry
 import com.swoval.files.Directory.Observer
 import com.swoval.files.Directory.OnChange
@@ -18,6 +19,7 @@ private[files] object Observers {
 
   /**
    * Simple observer that fires the same callback for all regular events and ignores any errors.
+   *
    * @param onchange The callback to fire when a file is created/updated/deleted
    * @tparam T The generic type of the [[Directory.Entry]]
    * @return An [[Observer]] instance
@@ -64,63 +66,70 @@ private[files] object Observers {
 /**
  * Container class that wraps multiple [[Observer]] and runs the callbacks for each whenever the
  * [[FileCache]] detects an event.
+ *
  * @tparam T The data type for the [[FileCache]] to which the observers correspond
  */
 private[files] class Observers[T] extends Observer[T] with AutoCloseable {
 
   private val counter: AtomicInteger = new AtomicInteger(0)
 
-  private val lock: AnyRef = new AnyRef()
+  private val lock: Lock = new Lock()
 
   private val observers: Map[Integer, Observer[T]] = new HashMap()
 
   override def onCreate(newEntry: Entry[T]): Unit = {
     var cbs: Collection[Observer[T]] = null
-    lock.synchronized {
-      cbs = observers.values
+    if (lock.lock()) {
+      try cbs = observers.values
+      finally lock.unlock()
+      val it: Iterator[Observer[T]] = cbs.iterator()
+      while (it.hasNext) it.next().onCreate(newEntry)
     }
-    val it: Iterator[Observer[T]] = cbs.iterator()
-    while (it.hasNext) it.next().onCreate(newEntry)
   }
 
   override def onDelete(oldEntry: Entry[T]): Unit = {
     var cbs: Collection[Observer[T]] = null
-    lock.synchronized {
-      cbs = observers.values
+    if (lock.lock()) {
+      try cbs = observers.values
+      finally lock.unlock()
+      val it: Iterator[Observer[T]] = cbs.iterator()
+      while (it.hasNext) it.next().onDelete(oldEntry)
     }
-    val it: Iterator[Observer[T]] = cbs.iterator()
-    while (it.hasNext) it.next().onDelete(oldEntry)
   }
 
   override def onUpdate(oldEntry: Entry[T], newEntry: Entry[T]): Unit = {
     var cbs: Collection[Observer[T]] = null
-    lock.synchronized {
-      cbs = observers.values
+    if (lock.lock()) {
+      try cbs = observers.values
+      finally lock.unlock()
+      val it: Iterator[Observer[T]] = cbs.iterator()
+      while (it.hasNext) it.next().onUpdate(oldEntry, newEntry)
     }
-    val it: Iterator[Observer[T]] = cbs.iterator()
-    while (it.hasNext) it.next().onUpdate(oldEntry, newEntry)
   }
 
   override def onError(path: Path, exception: IOException): Unit = {
     var cbs: Collection[Observer[T]] = null
-    lock.synchronized {
-      cbs = observers.values
+    if (lock.lock()) {
+      try cbs = observers.values
+      finally lock.unlock()
+      val it: Iterator[Observer[T]] = cbs.iterator()
+      while (it.hasNext) it.next().onError(path, exception)
     }
-    val it: Iterator[Observer[T]] = cbs.iterator()
-    while (it.hasNext) it.next().onError(path, exception)
   }
 
   def addObserver(observer: Observer[T]): Int = {
     val key: Int = counter.getAndIncrement
-    lock.synchronized {
-      observers.put(key, observer)
+    if (lock.lock()) {
+      try observers.put(key, observer)
+      finally lock.unlock()
     }
     key
   }
 
   def removeObserver(handle: Int): Unit = {
-    lock.synchronized {
-      observers.remove(handle)
+    if (lock.lock()) {
+      try observers.remove(handle)
+      finally lock.unlock()
     }
   }
