+++ files/jvm/src/main/java/com/swoval/files/Observers.java
@@ -1,5 +1,6 @@
 package com.swoval.files;
 
+import com.swoval.concurrent.Lock;
 import com.swoval.files.Directory.Entry;
 import com.swoval.files.Directory.Observer;
 import com.swoval.files.Directory.OnChange;
@@ -16,64 +17,89 @@ import java.util.concurrent.atomic.AtomicInteger;
 /**
  * Container class that wraps multiple {@link Observer} and runs the callbacks for each whenever the
  * {@link FileCache} detects an event.
+ *
  * @param <T> The data type for the {@link FileCache} to which the observers correspond
  */
 class Observers<T> implements Observer<T>, AutoCloseable {
   private final AtomicInteger counter = new AtomicInteger(0);
-  private final Object lock = new Object();
+  private final Lock lock = new Lock();
   private final Map<Integer, Observer<T>> observers = new HashMap<>();
 
   @Override
   public void onCreate(final Entry<T> newEntry) {
     final Collection<Observer<T>> cbs;
-    synchronized (lock) {
-      cbs = observers.values();
+    if (lock.lock()) {
+      try {
+        cbs = observers.values();
+      } finally {
+        lock.unlock();
+      }
+      final Iterator<Observer<T>> it = cbs.iterator();
+      while (it.hasNext()) it.next().onCreate(newEntry);
     }
-    final Iterator<Observer<T>> it = cbs.iterator();
-    while (it.hasNext()) it.next().onCreate(newEntry);
   }
 
   @Override
   public void onDelete(final Entry<T> oldEntry) {
     final Collection<Observer<T>> cbs;
-    synchronized (lock) {
-      cbs = observers.values();
+    if (lock.lock()) {
+      try {
+        cbs = observers.values();
+      } finally {
+        lock.unlock();
+      }
+      final Iterator<Observer<T>> it = cbs.iterator();
+      while (it.hasNext()) it.next().onDelete(oldEntry);
     }
-    final Iterator<Observer<T>> it = cbs.iterator();
-    while (it.hasNext()) it.next().onDelete(oldEntry);
   }
 
   @Override
   public void onUpdate(final Entry<T> oldEntry, final Entry<T> newEntry) {
     final Collection<Observer<T>> cbs;
-    synchronized (lock) {
-      cbs = observers.values();
+    if (lock.lock()) {
+      try {
+        cbs = observers.values();
+      } finally {
+        lock.unlock();
+      }
+      final Iterator<Observer<T>> it = cbs.iterator();
+      while (it.hasNext()) it.next().onUpdate(oldEntry, newEntry);
     }
-    final Iterator<Observer<T>> it = cbs.iterator();
-    while (it.hasNext()) it.next().onUpdate(oldEntry, newEntry);
   }
 
   @Override
   public void onError(Path path, IOException exception) {
     final Collection<Observer<T>> cbs;
-    synchronized (lock) {
-      cbs = observers.values();
+    if (lock.lock()) {
+      try {
+        cbs = observers.values();
+      } finally {
+        lock.unlock();
+      }
+      final Iterator<Observer<T>> it = cbs.iterator();
+      while (it.hasNext()) it.next().onError(path, exception);
     }
-    final Iterator<Observer<T>> it = cbs.iterator();
-    while (it.hasNext()) it.next().onError(path, exception);
   }
 
   public int addObserver(Observer<T> observer) {
     final int key = counter.getAndIncrement();
-    synchronized (lock) {
-      observers.put(key, observer);
+    if (lock.lock()) {
+      try {
+        observers.put(key, observer);
+      } finally {
+        lock.unlock();
+      }
     }
     return key;
   }
 
   public void removeObserver(int handle) {
-    synchronized (lock) {
-      observers.remove(handle);
+    if (lock.lock()) {
+      try {
+        observers.remove(handle);
+      } finally {
+        lock.unlock();
+      }
     }
   }
 
@@ -84,6 +110,7 @@ class Observers<T> implements Observer<T>, AutoCloseable {
 
   /**
    * Simple observer that fires the same callback for all regular events and ignores any errors.
+   *
    * @param onchange The callback to fire when a file is created/updated/deleted
    * @param <T> The generic type of the {@link Directory.Entry}
    * @return An {@link Observer} instance
@@ -106,13 +133,15 @@ class Observers<T> implements Observer<T>, AutoCloseable {
       }
 
       @Override
-      public void onError(final Path path, final IOException e) {
-      }
+      public void onError(final Path path, final IOException e) {}
     };
   }
 
   public static <T> Observer<T> apply(
-      final OnChange<T> oncreate, final OnUpdate<T> onupdate, final OnChange<T> ondelete, final OnError onerror) {
+      final OnChange<T> oncreate,
+      final OnUpdate<T> onupdate,
+      final OnChange<T> ondelete,
+      final OnError onerror) {
     return new Observer<T>() {
       @Override
       public void onCreate(final Entry<T> newEntry) {
